chore: make starknet_api_test_utils enabled in test mode
