+++ crates/gateway/Cargo.toml
@@ -13,8 +13,7 @@ testing = []
 
 [dependencies]
 axum.workspace = true
-assert_matches.workspace = true
-blockifier.workspace = true
+blockifier= { workspace = true , features = ["testing"] }
 cairo-lang-starknet-classes.workspace = true
 cairo-vm.workspace = true
 hyper.workspace = true
@@ -32,6 +31,7 @@ tokio.workspace = true
 validator.workspace = true
 
 [dev-dependencies]
+assert_matches.workspace = true
 pretty_assertions.workspace = true
 rstest.workspace = true
 starknet_mempool = { path = "../mempool", version = "0.0" }
