+++ crates/gateway/src/lib.rs
@@ -4,7 +4,6 @@ pub mod errors;
 pub mod gateway;
 pub mod rpc_objects;
 pub mod rpc_state_reader;
-pub mod starknet_api_test_utils;
 pub mod state_reader;
 #[cfg(test)]
 pub mod state_reader_test_utils;
