+++ crates/gateway/src/stateless_transaction_validator_test.rs
@@ -4,14 +4,14 @@ use starknet_api::hash::StarkFelt;
 use starknet_api::rpc_transaction::{ContractClass, ResourceBoundsMapping};
 use starknet_api::transaction::{Calldata, Resource, ResourceBounds, TransactionSignature};
 use starknet_api::{calldata, stark_felt};
-
-use crate::compiler_version::{VersionId, VersionIdError};
-use crate::config::StatelessTransactionValidatorConfig;
-use crate::declare_tx_args;
-use crate::starknet_api_test_utils::{
+use test_utils::declare_tx_args;
+use test_utils::starknet_api_test_utils::{
     create_resource_bounds_mapping, external_declare_tx, external_tx_for_testing,
     zero_resource_bounds_mapping, TransactionType, NON_EMPTY_RESOURCE_BOUNDS,
 };
+
+use crate::compiler_version::{VersionId, VersionIdError};
+use crate::config::StatelessTransactionValidatorConfig;
 use crate::stateless_transaction_validator::{
     StatelessTransactionValidator, StatelessTransactionValidatorError,
 };
