+++ crates/test_utils/Cargo.toml
@@ -8,5 +8,9 @@ license.workspace = true
 [features]
 
 [dependencies]
+assert_matches.workspace = true
+blockifier.workspace = true
+starknet_api.workspace = true
+serde_json.workspace = true
