+++ crates/test_utils/src/lib.rs
@@ -1,6 +1,8 @@
 use std::env;
 use std::path::{Path, PathBuf};
 
+pub mod starknet_api_test_utils;
+
 pub const TEST_FILES_FOLDER: &str = "crates/test_utils/test_files";
 pub const CONTRACT_CLASS_FILE: &str = "contract_class.json";
 pub const COMPILED_CLASS_HASH_OF_CONTRACT_CLASS: &str =
